properly display token data

+++ src/ui/pages/OpNet/Mint.tsx
@@ -250,7 +250,7 @@ export default function Mint() {
                 isToSign: false,
                 opneTokens: [
                   {
-                    amount: parseInt(inputAmount) * 10 ** OpNetBalance.divisibility,
+                    amount: parseFloat(inputAmount) * 10 ** OpNetBalance.divisibility,
                     divisibility: OpNetBalance.divisibility,
                     spacedRune: OpNetBalance.name,
                     symbol: OpNetBalance.symbol
